+++ crates/voicevox_core/examples/talk.rs
@@ -52,6 +52,8 @@ struct Args {
 }
 
 fn main() -> anyhow::Result<()> {
+    tracing_subscriber::fmt().init(); // `voicevox_core`や(VOICEVOX) ONNX Runtimeの情報表示や警告は、`tracing`に出力される。
+
     let args = Args::parse();
 
     // ONNX Runtimeのロード