+++ crates/voicevox_core/src/engine/talk/full_context_label.rs
@@ -2,10 +2,14 @@ use std::str::FromStr;
 
 use jlabel::Label;
 use smallvec::SmallVec;
+use typed_floats::tf32;
 
 use crate::AccentPhrase;
 
-use super::{super::mora_mappings::MORA_PHONEMES_TO_MORA_KANA, open_jtalk::FullcontextExtractor};
+use super::{
+    super::{acoustic_feature_extractor::NonConsonant, mora_mappings::MORA_PHONEMES_TO_MORA_KANA},
+    open_jtalk::FullcontextExtractor,
+};
 
 #[derive(thiserror::Error, Debug)]
 #[error("入力テキストからのフルコンテキストラベル抽出に失敗しました: {context}")]
@@ -23,12 +27,21 @@ enum ErrorKind {
     #[display("jlabelでラベルを解釈することができませんでした")]
     Jlabel,
 
+    #[display("VOICEVOXの`consonant`として不正、もしくは未知の音素が発生しました: {_0}")]
+    InvalidConsonant(String),
+
     #[display("too long mora")]
     TooLongMora,
 }
 
 type Result<T> = std::result::Result<T, FullContextLabelError>;
 
+/// # Panics
+///
+/// 解析結果が次の場合にパニックする。
+///
+/// - アクセント位置として`0`が存在する。
+/// - [`generate_moras`]のパニック条件を満たす。
 pub(crate) fn extract_full_context_label(
     open_jtalk: &impl FullcontextExtractor,
     text: impl AsRef<str>,
@@ -56,6 +69,10 @@ pub(crate) fn extract_full_context_label(
     })
 }
 
+/// # Panics
+///
+/// - アクセント位置として`0`が存在する。
+/// - [`generate_moras`]のパニック条件を満たす。
 fn generate_accent_phrases(
     utterance: &[Label],
 ) -> std::result::Result<Vec<AccentPhrase>, ErrorKind> {
@@ -92,9 +109,9 @@ fn generate_accent_phrases(
                 text: "、".into(),
                 consonant: None,
                 consonant_length: None,
-                vowel: "pau".into(),
-                vowel_length: 0.,
-                pitch: 0.,
+                vowel: NonConsonant::MorablePau,
+                vowel_length: tf32::ZERO,
+                pitch: tf32::ZERO.into(),
             })
         } else {
             None
@@ -102,6 +119,10 @@ fn generate_accent_phrases(
 
         // workaround for VOICEVOX/voicevox_engine#55
         let accent = usize::from(ap_curr.accent_position).min(moras.len());
+
+        let accent = accent
+            .try_into()
+            .expect("Open JTalk is not considered to emit `0` as of v1.11");
 
         accent_phrases.push(AccentPhrase {
             moras,
@@ -113,6 +134,9 @@ fn generate_accent_phrases(
     Ok(accent_phrases)
 }
 
+/// # Panics
+///
+/// 母音にあたる[`Label`]の[`jlabel::Phoneme::c`]が`None`か、あるいは[`NonConsonant`]にあてはまらなければパニックする。
 fn generate_moras(accent_phrase: &[Label]) -> std::result::Result<Vec<crate::Mora>, ErrorKind> {
     let mut moras = Vec::with_capacity(accent_phrase.len());
 
@@ -122,11 +146,11 @@ fn generate_moras(accent_phrase: &[Label]) -> std::result::Result<Vec<crate::Mor
             labels.iter().filter(|label| label.mora.is_some()).collect();
         match labels[..] {
             [consonant, vowel] => {
-                let mora = generate_mora(Some(consonant), vowel);
+                let mora = generate_mora(Some(consonant), vowel)?;
                 moras.push(mora);
             }
             [vowel] => {
-                let mora = generate_mora(None, vowel);
+                let mora = generate_mora(None, vowel)?;
                 moras.push(mora);
             }
             // silやpau以外の音素がないモーラは含めない
@@ -154,17 +178,34 @@ fn generate_moras(accent_phrase: &[Label]) -> std::result::Result<Vec<crate::Mor
     Ok(moras)
 }
 
-fn generate_mora(consonant: Option<&Label>, vowel: &Label) -> crate::Mora {
-    let consonant_phoneme = consonant.and_then(|c| c.phoneme.c.to_owned());
-    let vowel = vowel.phoneme.c.clone().unwrap();
-    crate::Mora {
-        text: mora_to_text(consonant_phoneme.as_deref(), &vowel),
-        consonant: consonant_phoneme,
-        consonant_length: consonant.and(Some(0.0)),
-        vowel,
-        vowel_length: 0.0,
-        pitch: 0.0,
-    }
+/// # Panics
+///
+/// `vowel.phoneme.c`が`None`か、あるいは[`NonConsonant`]にあてはまらなければパニックする。
+fn generate_mora(
+    consonant: Option<&Label>,
+    vowel: &Label,
+) -> std::result::Result<crate::Mora, ErrorKind> {
+    let consonant = consonant.and_then(|c| c.phoneme.c.as_deref());
+    let vowel = vowel.phoneme.c.as_deref().unwrap();
+    Ok(crate::Mora {
+        text: mora_to_text(consonant, vowel),
+        consonant: consonant
+            .map(|consonant| {
+                consonant.parse().map_err(|_| {
+                    // VOICEVOX/voicevox_engine#57 への対策が無いため、
+                    // 例えば`"ア".repeat(99)`を入力すると溢れた`a`が子音として入ってきてしまうらしい。
+                    // FIXME: さらに調査した上 VOICEVOX/voicevox_engine#58 と同じ方向で、アクセント位置の`49`の壁の対策をする。
+                    ErrorKind::InvalidConsonant(consonant.to_owned())
+                })
+            })
+            .transpose()?,
+        consonant_length: consonant.and(Some(tf32::ZERO)),
+        vowel: vowel
+            .parse()
+            .expect("Open JTalk is not considered to emit consonants here as of v1.11"),
+        vowel_length: tf32::ZERO,
+        pitch: tf32::ZERO.into(),
+    })
 }
 
 pub fn mora_to_text(consonant: Option<&str>, vowel: &str) -> String {
@@ -196,8 +237,9 @@ mod tests {
     use pretty_assertions::assert_eq;
     use rstest::rstest;
     use rstest_reuse::*;
+    use typed_floats::tf32;
 
-    use crate::AccentPhrase;
+    use crate::{AccentPhrase, numerics::non_zero};
 
     use super::super::{
         Mora,
@@ -209,11 +251,11 @@ mod tests {
     fn mora(text: &str, consonant: Option<&str>, vowel: &str) -> Mora {
         Mora {
             text: text.into(),
-            consonant: consonant.map(|c| c.into()),
-            consonant_length: consonant.and(Some(0.0)),
-            vowel: vowel.into(),
-            vowel_length: 0.0,
-            pitch: 0.0,
+            consonant: consonant.map(|c| c.parse().unwrap()),
+            consonant_length: consonant.and(Some(tf32::ZERO)),
+            vowel: vowel.parse().unwrap(),
+            vowel_length: tf32::ZERO,
+            pitch: tf32::ZERO.into(),
         }
     }
 
@@ -230,7 +272,7 @@ mod tests {
         &[
             AccentPhrase {
                 moras: vec![mora("イェ", Some("y"), "e")],
-                accent: 1,
+                accent: non_zero!(1usize),
                 pause_mora: None,
                 is_interrogative: false,
             }
@@ -252,7 +294,7 @@ mod tests {
                     mora("ン", None, "N"),
                     mora("ッ", None, "cl"),
                 ],
-                accent: 3,
+                accent: non_zero!(3usize),
                 pause_mora: None,
                 is_interrogative: false,
             },
@@ -287,7 +329,7 @@ mod tests {
                     mora("レ", Some("r"), "e"),
                     mora("ワ", Some("w"), "a"),
                 ],
-                accent: 3,
+                accent: non_zero!(3usize),
                 pause_mora: None,
                 is_interrogative: false,
             },
@@ -299,7 +341,7 @@ mod tests {
                     mora("デ", Some("d"), "e"),
                     mora("ス", Some("s"), "U"),
                 ],
-                accent: 1,
+                accent: non_zero!(1usize),
                 pause_mora: None,
                 is_interrogative: false,
             },
@@ -339,7 +381,7 @@ mod tests {
                     mora("イ", None, "i"),
                     mora("チ", Some("ch"), "i"),
                 ],
-                accent: 2,
+                accent: non_zero!(2usize),
                 pause_mora: Some(mora("、", None, "pau")),
                 is_interrogative: false,
             },
@@ -348,7 +390,7 @@ mod tests {
                     mora("セ", Some("s"), "e"),
                     mora("ン", None, "N"),
                 ],
-                accent: 1,
+                accent: non_zero!(1usize),
                 pause_mora: Some(mora("、", None, "pau")),
                 is_interrogative: false,
             },
@@ -359,7 +401,7 @@ mod tests {
                     mora("マ", Some("m"), "a"),
                     mora("ン", None, "N"),
                 ],
-                accent: 3,
+                accent: non_zero!(3usize),
                 pause_mora: Some(mora("、", None, "pau")),
                 is_interrogative: false,
             },
@@ -370,7 +412,7 @@ mod tests {
                     mora("オ", None, "o"),
                     mora("ク", Some("k"), "u"),
                 ],
-                accent: 2,
+                accent: non_zero!(2usize),
                 pause_mora: None,
                 is_interrogative: true,
             },
@@ -404,7 +446,7 @@ mod tests {
                     mora("ッ", None, "cl"),
                     mora("ト", Some("t"), "o"),
                 ],
-                accent: 3,
+                accent: non_zero!(3usize),
                 pause_mora: Some(mora("、", None, "pau")),
                 is_interrogative: false,
             },
@@ -413,13 +455,13 @@ mod tests {
                     mora("ア", None, "a"),
                     mora("ア", None, "a"),
                 ],
-                accent: 1,
+                accent: non_zero!(1usize),
                 pause_mora: None,
                 is_interrogative: false,
             },
             AccentPhrase {
                 moras: vec![mora("ア", None, "a")],
-                accent: 1,
+                accent: non_zero!(1usize),
                 pause_mora: None,
                 is_interrogative: false,
             },