+++ CHANGELOG.md
@@ -42,8 +42,9 @@
 ### Added
 
 - `Synthesizer::load_voice_model`にオプション`on_existing`が追加されます ([#1331], [#1337])。
-- \[Rust\] \[Linux\] APIドキュメントが改善されます ([#1343])。
-    - muslターゲットでは`load-onnxruntime`が事実上利用不可であることが明記されます。
+- \[Rust\] APIドキュメントが改善されます ([#1343], [#1381])。
+    - トップページのコード例で`tracing_subscriber::fmt().init();`が行われるようになります。
+    - \[Linux\] muslターゲットでは`load-onnxruntime`が事実上利用不可であることが明記されます。
 - \[C\] \[macOS\] :tada: macOS向けのXCFrameworkがvoicevox\_core-xcframework-cpu-{バージョン}.zipという名前でリリースされるようになります ([#1056] helped by [@nekomimimi], [#1114], [#1362])。
 - \[C,ダウンローダー\] \[macOS\] リリースがコード署名されるようになります ([#1326])。
 - \[ダウンローダー\] `--os`オプションで`android`と`ios`を指定できるようになります。ただしiOSの`c-api`をダウンロードすることはできません ([#1313])。
@@ -51,6 +52,7 @@
 
 ### Changed
 
+- \[BREAKING\] `AudioQuery`/`AccentPhrase`/`Mora`の警告が、`output_sampling_rate`のものを除きエラーになります。また入力するテキストにより`AccentPhrase`作成時点でエラーになるケースがあります。Rust APIにおいては、制約が強くなる形で各フィールドの型が変わります ([#1384])。
 - \[Rust\] \[BREAKING\] `Synthesizer::load_voice_model`がビルダースタイルになります ([#1331])。
 - \[Rust\] \[BREAKING\] MSRVが1.89.0になります ([#1323])。
 - \[Rust\] \[BREAKING\] `load-onnxruntime`フィーチャと`link-onnxruntime`フィーチャの両方において、ビルド時のダウンロードおよびリンカーフラグの設定が[pykeio/ort](https://github.com/pykeio/ort)由来の処理に依存しなくなります。それにより、ビルド時の挙動が以下の点で変わります。なお以下で言及する"ONNX Runtime"はVOICEVOX ONNX Runtimeとは異なることに注意してください [#1278]。
@@ -74,7 +76,9 @@
 
 ### Fixed
 
+- 一部のケースにおける`RunModel`エラーのメッセージが改善されます ([#1385])。
 - \[Python\] `Synthesizer.load_voice_model`のdocstringにおける"Parameters"の説明が誤っていたのが修正されます ([#1359])。
+- \[Java\] Java APIの実装当初から事実上利用不可だった`Synthesizer.Builder#cpuNumThreads`が利用可能になります ([#1394])。
 
 ### Security
 
@@ -1516,6 +1520,10 @@ Windows版ダウンローダーのビルドに失敗しています。
 [#1350]: https://github.com/VOICEVOX/voicevox_core/pull/1350
 [#1359]: https://github.com/VOICEVOX/voicevox_core/pull/1359
 [#1362]: https://github.com/VOICEVOX/voicevox_core/pull/1362
+[#1381]: https://github.com/VOICEVOX/voicevox_core/pull/1381
+[#1384]: https://github.com/VOICEVOX/voicevox_core/pull/1384
+[#1385]: https://github.com/VOICEVOX/voicevox_core/pull/1385
+[#1394]: https://github.com/VOICEVOX/voicevox_core/pull/1394
 
 [VOICEVOX/onnxruntime-builder#25]: https://github.com/VOICEVOX/onnxruntime-builder/pull/25